refactored from app component to dedicated component
