+++ src/app/app-routing.module.ts
@@ -1,10 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { CalendarViewComponent } from './container/calendar-view/calendar-view.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'calendar', component: CalendarViewComponent },
+  { path: '**', redirectTo: 'calendar' },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
